+++ changelog.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 1.4.1 - 2022-03-17
+
+### Added
+
+- Added a new `LinkToInstance` method which will instead use `Instance.Destroying`.
+- Added traceback to `Janitor:AddPromise` for invalid promises.
+
+### Changed
+
+- The legacy `LinkToInstance` method has been renamed to `LegacyLinkToInstance`.
+
+### Fixed
+
+- Fixed Janitor not warning about an invalid `MethodName` for threads and functions.
+- Fixed incorrect documentation about `Janitor.CurrentlyCleaning`.
+
 ## 1.14 - 2022-03-12
 
 ### Added